Create 'deploy' stage with key for matyasselmeci/htcondor-ce

+++ tests/setup_tests.sh
@@ -6,11 +6,16 @@
 # We use `--privileged` for cgroup compatability, which seems to be enabled by default in HTCondor 8.6.x
 if [ "${OS_VERSION}" -eq 6 ]; then
 
-    sudo docker run --privileged --rm=true \
+    if ${DEPLOY_STAGE}; then
+        RM=false
+    else
+        RM=true
+    fi
+    sudo docker run --privileged --rm=$RM \
          --volume /sys/fs/cgroup:/sys/fs/cgroup \
          --volume `pwd`:/htcondor-ce:rw \
          centos:centos${OS_VERSION} \
-         /bin/bash -c "bash -xe /htcondor-ce/tests/test_inside_docker.sh ${OS_VERSION} ${BUILD_ENV}"
+         /bin/bash -c "bash -xe /htcondor-ce/tests/test_inside_docker.sh ${OS_VERSION} ${BUILD_ENV} ${DEPLOY_STAGE}"
 
 elif [ "${OS_VERSION}" -eq 7 ]; then
 
@@ -23,13 +28,14 @@ elif [ "${OS_VERSION}" -eq 7 ]; then
     DOCKER_CONTAINER_ID=$(docker ps | grep centos | awk '{print $1}')
     docker logs $DOCKER_CONTAINER_ID
     docker exec --tty --interactive $DOCKER_CONTAINER_ID \
-           /bin/bash -xec "bash -xe /htcondor-ce/tests/test_inside_docker.sh ${OS_VERSION} ${BUILD_ENV};
+           /bin/bash -xec "bash -xe /htcondor-ce/tests/test_inside_docker.sh ${OS_VERSION} ${BUILD_ENV} ${DEPLOY_STAGE};
            echo -ne \"------\nEND HTCONDOR-CE TESTS\n\";"
 
-    docker ps -a
-    docker stop $DOCKER_CONTAINER_ID
-    docker rm -v $DOCKER_CONTAINER_ID
-
+        docker ps -a
+        docker stop $DOCKER_CONTAINER_ID
+    if ! ${DEPLOY_STAGE}; then
+        docker rm -v $DOCKER_CONTAINER_ID
+    fi
 fi
